+++ NEWS.md
@@ -9,6 +9,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Changed
 
+* `create_ws_from_data` has new arguments to personnalize the workspace: context, sap_name and path
 * `create_insee_regressors` gives new names for calendar regressors : "REG2_week" instead of "REG2_group1" etc.
 * `init_env()` creates a project with DESCRIPTION file with `Type: Project` and {rjd3qr} as dependency  [#83](https://github.com/InseeFr/rjd3production/issues/83)